+++ src/components/DatePicker/Calendar.tsx
@@ -1,16 +1,12 @@
-import { ComponentPropsWithoutRef, PropsWithChildren } from "react";
-
 import dayjs, { type Dayjs } from "dayjs";
 
+import {
+  CalendarItemProps,
+  CalendarProps,
+  daysOfTheWeek,
+} from "@/constants/calendar";
+import { getWeeks } from "@/utils/calendar";
 import { cn } from "@/utils/cn";
-
-interface CalendarItemProps
-  extends PropsWithChildren,
-    ComponentPropsWithoutRef<"div"> {
-  dayOfTheWeek: number;
-  isInactive?: boolean;
-  isSelected?: boolean;
-}
 
 const CalendarItem = ({
   children,
@@ -39,8 +35,6 @@ const CalendarItem = ({
   );
 };
 
-const daysOfTheWeek = ["일", "월", "화", "수", "목", "금", "토"];
-
 const CalendarHead = () => (
   <div className="flex items-center justify-around">
     {daysOfTheWeek.map((dayOfTheWeek, i) => (
@@ -55,14 +49,6 @@ const CalendarHead = () => (
   </div>
 );
 
-interface CalendarProps {
-  firstDayOfMonth: Dayjs;
-  selectedDate?: Date;
-  limit?: number;
-  onChangeSelectedDate: (day: Date) => void;
-  onChangeFirstDayOfMonth: (day: Dayjs) => void;
-}
-
 const Calendar = ({
   firstDayOfMonth,
   selectedDate,
@@ -70,33 +56,6 @@ const Calendar = ({
   onChangeSelectedDate,
   onChangeFirstDayOfMonth,
 }: CalendarProps) => {
-  const getWeeks = () => {
-    const daysInMonth = firstDayOfMonth.daysInMonth();
-    const lastDayOfMonth = firstDayOfMonth.date(daysInMonth);
-
-    const previousMonthDays = Array.from(
-      { length: firstDayOfMonth.day() },
-      (_, i) => firstDayOfMonth.date(firstDayOfMonth.day() * -1 + i + 1),
-    );
-
-    const currentMonthDays = Array.from({ length: daysInMonth }, (_, i) =>
-      firstDayOfMonth.date(i + 1),
-    );
-
-    const nextMonthDays = Array.from(
-      { length: 6 - lastDayOfMonth.day() },
-      (_, i) => firstDayOfMonth.date(daysInMonth + i + 1),
-    );
-
-    const days = [...previousMonthDays, ...currentMonthDays, ...nextMonthDays];
-
-    const weeks = Array.from({ length: days.length / 7 }, (_, i) => [
-      ...days.slice(i * 7, (i + 1) * 7),
-    ]);
-
-    return weeks;
-  };
-
   const handleClickDate = (day: Dayjs, isInactive: boolean) => {
     if (isInactive) return;
 
@@ -104,13 +63,13 @@ const Calendar = ({
       onChangeFirstDayOfMonth(day.date(1));
     }
 
-    onChangeSelectedDate(day.toDate());
+    onChangeSelectedDate(dayjs(day));
   };
 
   return (
     <div className="flex w-full flex-col gap-2 text-black">
       <CalendarHead />
-      {getWeeks().map(week => (
+      {getWeeks(firstDayOfMonth).map(week => (
         <div
           key={week[0].valueOf()}
           className="flex items-center justify-around"
