+++ src/constants/calendar.ts
@@ -1,0 +1,22 @@
+import { ComponentPropsWithoutRef, PropsWithChildren } from "react";
+
+import { Dayjs } from "dayjs";
+
+export const daysOfTheWeek = ["일", "월", "화", "수", "목", "금", "토"];
+
+export interface CalendarProps {
+  firstDayOfMonth: Dayjs;
+  selectedDate?: Dayjs;
+  limit?: number;
+  onChangeSelectedDate: (day: Dayjs) => void;
+  onChangeFirstDayOfMonth: (day: Dayjs) => void;
+}
+
+export interface CalendarItemProps
+  extends PropsWithChildren,
+    ComponentPropsWithoutRef<"div"> {
+  dayOfTheWeek: number;
+  type?: string;
+  isInactive?: boolean;
+  isSelected?: boolean;
+}
